requirements.txt
PyYAML==6.0.2
Flask==3.1.0
